convert AoEForm to TS
